Implement sign in page auth state based navigation
Attached the onAuthStateChanged observer to the FIREBASE_AUTH
object to trigger a callback function every time auth state changes.

If the auth state changes and FIREBASE_AUTH has a user object,
then the user is navigated to the main page inside the app, otherwise
nothing happens (since they are already on the sign in page).

TODO: Testing still needs to be done to see if the user is automatically
navigated to the main page inside the app upon a successful
registration.

TODO: Navigate the user from inside the app to outside the app upon
clicking a log out button or the auth state changing.