Enable log out navigation to sign in page and auth sign out
I was able to get the log out button on the "one" page (to be
called "home" in the future) to sign the user out and navigate
them back to the sign in screen.

BUGS: Whenever user logs in or logs out, the onAuthStateChange hook
is run one more time than it was previously. Functionally this has
no impact other than redundant logs, but could pose an issue in the
future.

TODO (1): Test to make sure auth state is NOT changed upon user
registration. Upon reflection, I realized changing auth state after
registering a user would present problems for when future versions
implement email verification. The most that should be done after a user
registers is that the yare brought back to the sign in page.

TODO (2): Make auth state presistent. Likely going to use
react-native-async-storage, but I have had issues with the imports
being recognized. May be a Typescript problem, or problem with the
package itself.