Merge auth-based-nav into stage (#14)
* Implement sign in page auth state based navigation

Attached the onAuthStateChanged observer to the FIREBASE_AUTH
object to trigger a callback function every time auth state changes.

If the auth state changes and FIREBASE_AUTH has a user object,
then the user is navigated to the main page inside the app, otherwise
nothing happens (since they are already on the sign in page).

TODO: Testing still needs to be done to see if the user is automatically
navigated to the main page inside the app upon a successful
registration.

TODO: Navigate the user from inside the app to outside the app upon
clicking a log out button or the auth state changing.

* Enable log out navigation to sign in page and auth sign out

I was able to get the log out button on the "one" page (to be
called "home" in the future) to sign the user out and navigate
them back to the sign in screen.

BUGS: Whenever user logs in or logs out, the onAuthStateChange hook
is run one more time than it was previously. Functionally this has
no impact other than redundant logs, but could pose an issue in the
future.

TODO (1): Test to make sure auth state is NOT changed upon user
registration. Upon reflection, I realized changing auth state after
registering a user would present problems for when future versions
implement email verification. The most that should be done after a user
registers is that the yare brought back to the sign in page.

TODO (2): Make auth state presistent. Likely going to use
react-native-async-storage, but I have had issues with the imports
being recognized. May be a Typescript problem, or problem with the
package itself.

* Merge home-page into auth-based-nav (#13)

* Rename pageone and pagetwo to Home and Profile pages

* Fix capitalization of page names

* Add drawer based navigation and more pages

-Replaced tabs with a drawer
-Added Games and Rewards skeleton pages

* Change to a more conventional navigation order

---------

Co-authored-by: brennamehl <brennam@email.unc.edu>
Co-authored-by: Moh88CS <mohamed.abdulal88@gmail.com>

---------

Co-authored-by: brennamehl <brennam@email.unc.edu>
Co-authored-by: Moh88CS <mohamed.abdulal88@gmail.com>